pipeline-azure/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
scikit-learn>=1.3.0
pandas>=2.0.0
pyarrow>=14.0.0
joblib>=1.3.0
